+++ mDone/App/AppDependencies.swift
@@ -12,7 +12,6 @@ struct AppDependencies {
             CachedLabel.self,
             PendingOperation.self,
             FocusRecord.self,
-            TaskEstimate.self,
         ])
 
         let config = ModelConfiguration(schema: schema, isStoredInMemoryOnly: false)